requirements.txt
pyrogram
pillow
easyocr
python-dotenv